adding selenium
