+++ pom.xml
@@ -207,6 +207,11 @@
                     </properties>
                 </configuration>
             </plugin>
+            <plugin>
+                <groupId>external.atlassian.jgitflow</groupId>
+                <artifactId>jgitflow-maven-plugin</artifactId>
+                <version>1.0-m5.1</version>
+            </plugin>
         </plugins>
     </build>
 </project>
